requirements.txt
streamlit>=1.24.0
supabase>=1.0.3
python-dotenv>=1.0.0